# Conflicts:
#	src/main/java/com/honeybee/controller/MypageController.java
#	src/main/java/com/honeybee/domain/AttachFileDTO.java
#	src/main/java/com/honeybee/service/FreeReplyService.java
#	src/main/java/com/honeybee/service/FreeReplyServiceImpl.java
#	src/main/java/com/honeybee/service/FreeService.java
#	src/main/webapp/WEB-INF/views/free/get.jsp

+++ src/main/java/com/honeybee/domain/FreeReplyVO.java
@@ -17,7 +17,7 @@ public class FreeReplyVO { // 자유게시물 댓글
 	
 	private int thumb; // 댓글 좋아요 수
 	private int layer; // 댓글 깊이
-	private int bundle; // 댓글 묶음
+	private Long bundle; // 댓글 묶음
 	private double bunorder; // 댓글 묶음 순서
 	
 	private Date regdt; // 작성일자
